+++ RELEASE-NOTES.md
@@ -69,6 +69,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Inlet: Renamed `Table` to `Container`
 - Inlet collections of mixed or incorrect type will now fail verification, even if they're
   not marked as required
+- Required collections no longer fail Inlet verification if they are empty in the input file
 
 ### Fixed
 - Updated to new BLT version that does not fail when ClangFormat returns an empty